[DEV16] Nouveau composant Angular pour le tri des notes
