+++ src/app/shared/_models/note-searching-data.model.ts
